backend/requirements.txt
fastapi
uvicorn
faiss-cpu
google-genai
pdfplumber
pytesseract
pdf2image
pillow
requests